*** empty log message ***

+++ src/solv_simple/fn_xor/bt_specfn_xor.cc
@@ -54,7 +54,7 @@ int ApplyInferenceToXOR(int nBranchVar, bool bBVPolarity, int nSmurfNumber, void
 	arrSmurfStates[nSmurfNumber] = pTrueSimpleSmurfState;
 	SimpleSmurfProblemState->arrSmurfStack[SimpleSmurfProblemState->nCurrSearchTreeLevel].nNumSmurfsSatisfied++;
 	
-	d7_printf3("      XORSmurf %d transitioned to state %x\n", nSmurfNumber, (unsigned int)arrSmurfStates[nSmurfNumber]);
+	d7_printf3("      XORSmurf %d transitioned to state %x\n", nSmurfNumber, arrSmurfStates[nSmurfNumber]);
 	return 1;
 }
 
@@ -88,6 +88,6 @@ int ApplyInferenceToXORCounter(int nBranchVar, bool bBVPolarity, int nSmurfNumbe
 //	if(arrSmurfStates[nSmurfNumber] == pTrueSimpleSmurfState) //Can't happen here
 //	  SimpleSmurfProblemState->arrSmurfStack[SimpleSmurfProblemState->nCurrSearchTreeLevel].nNumSmurfsSatisfied++;
 	
-	d7_printf3("      XORCounterSmurf %d transitioned to state %x\n", nSmurfNumber, (unsigned int)arrSmurfStates[nSmurfNumber]);
+	d7_printf3("      XORCounterSmurf %d transitioned to state %x\n", nSmurfNumber, arrSmurfStates[nSmurfNumber]);
 	return 1;
 }
